refactor: move via stuff out to a separate class
Avoids blowing things up when ViaVersion isn't present.
