+++ platform/platform-sportpaper/src/main/java/dev/pgm/community/platform/sportpaper/features/serverlinks/ViaServerLinks.java
@@ -1,6 +1,4 @@
-package dev.pgm.community.platform.sportpaper.features;
-
-import static dev.pgm.community.util.Supports.Variant.SPORTPAPER;
+package dev.pgm.community.platform.sportpaper.features.serverlinks;
 
 import com.viaversion.nbt.tag.Tag;
 import com.viaversion.viaversion.api.Via;
@@ -12,49 +10,29 @@ import com.viaversion.viaversion.api.protocol.version.ProtocolVersion;
 import com.viaversion.viaversion.api.type.Types;
 import com.viaversion.viaversion.libs.gson.JsonParser;
 import com.viaversion.viaversion.libs.mcstructs.text.utils.JsonNbtConverter;
-import dev.pgm.community.serverlinks.ServerLinksFeature;
 import dev.pgm.community.serverlinks.types.ServerLink;
-import dev.pgm.community.util.Supports;
 import java.util.List;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.serializer.gson.GsonComponentSerializer;
 import org.bukkit.entity.Player;
 
-@Supports(SPORTPAPER)
-public class SpServerLinksPlatform implements ServerLinksFeature.ServerLinksPlatform {
-  private final Protocol<?, ?, ?, ?> serverLinkProtocol = findServerLinkProtocol();
-  private static final boolean hasVia = hasVia();
+public class ViaServerLinks {
+  private static final Protocol<?, ?, ?, ?> serverLinkProtocol = findServerLinkProtocol();
 
-  private static boolean hasVia() {
-    try {
-      Class.forName("com.viaversion.viaversion.api.Via");
-      return true;
-    } catch (ClassNotFoundException e) {
-      return false;
+  public static void sendToPlayer(Player player, List<ServerLink> serverLinks) {
+    if (!Via.getAPI().isInjected(player.getUniqueId())) return;
+    UserConnection userConnection = Via.getAPI().getConnection(player.getUniqueId());
+    if (userConnection != null
+        && userConnection
+            .getProtocolInfo()
+            .protocolVersion()
+            .newerThanOrEqualTo(ProtocolVersion.v1_21)) {
+      PacketWrapper serverLinksPacket = createPacket(userConnection, serverLinks);
+      serverLinksPacket.scheduleSend(serverLinkProtocol.getClass());
     }
   }
 
-  @Override
-  public boolean isSupported() {
-    return hasVia;
-  }
-
-  @Override
-  public void sendToPlayer(Player player, List<ServerLink> serverLinks) {
-    if (hasVia && Via.getAPI().isInjected(player.getUniqueId())) {
-      UserConnection userConnection = Via.getAPI().getConnection(player.getUniqueId());
-      if (userConnection != null
-          && userConnection
-              .getProtocolInfo()
-              .protocolVersion()
-              .newerThanOrEqualTo(ProtocolVersion.v1_21)) {
-        PacketWrapper serverLinksPacket = createPacket(userConnection, serverLinks);
-        serverLinksPacket.scheduleSend(serverLinkProtocol.getClass());
-      }
-    }
-  }
-
-  private PacketWrapper createPacket(UserConnection conn, List<ServerLink> links) {
+  private static PacketWrapper createPacket(UserConnection conn, List<ServerLink> links) {
     var packetTypes = serverLinkProtocol.getPacketTypesProvider().mappedClientboundPacketTypes();
     var packetType = packetTypes.get(State.PLAY).typeByName("SERVER_LINKS");
     PacketWrapper packet = PacketWrapper.create(packetType, conn);
@@ -73,12 +51,12 @@ public class SpServerLinksPlatform implements ServerLinksFeature.ServerLinksPlat
     return packet;
   }
 
-  private Tag toViaTag(Component component) {
+  private static Tag toViaTag(Component component) {
     return JsonNbtConverter.toNbt(
         JsonParser.parseString(GsonComponentSerializer.gson().serialize(component)));
   }
 
-  private Protocol<?, ?, ?, ?> findServerLinkProtocol() {
+  private static Protocol<?, ?, ?, ?> findServerLinkProtocol() {
     return Via.getManager()
         .getProtocolManager()
         .getProtocol(/* to */ ProtocolVersion.v1_21, /* from */ ProtocolVersion.v1_20_5);
